WIP - rewrite of Stackage-dist using discussed Stackable Roots feature

+++ make-stackage-dist.sh
@@ -2,14 +2,12 @@
 
 set -u
 
-RESOLVER=lts-5.0
-
 t=${BASH_SOURCE[0]}
 t=$(dirname ${t})
 t=$(realpath ${t})
 
 syntax() {
-    echo "build-srpm -t [type] -o [output directory] -n"
+    echo "build-srpm -t [resolver] -t [type] -o [output directory] -n"
 }
 
 outdir=""
@@ -18,10 +16,13 @@ pkgtype=
 maintainer=
 distros=
 
-while getopts "o:nt:m:d:" o; do
+while getopts "o:r:nt:m:d:" o; do
     case "${o}" in
         o)
             outdir=${OPTARG}
+            ;;
+        r)
+            resolver=${OPTARG}
             ;;
         t)
             pkgtype=${OPTARG}
@@ -69,64 +70,21 @@ case $pkgtype in
 	exit -1
 esac
 
-export STACK_ROOT=${t}/stackroot
+stack exec -- stack --help | grep relocate > /dev/null
+if [[ "$?" != 0 ]] ; then
+    echo "Need a newer Stack"
+    exit -1
+fi
+
+export STACK_ROOT=${t}/stack-root
 
 if [[ "$nocleanup" == "0" ]]; then
     rm -rf ${STACK_ROOT}
     mkdir -p ${STACK_ROOT}
 fi
 
-cat << EOF > ${STACK_ROOT}/config.yaml
-download-cache-paths:
-- ${STACK_ROOT}/download-cache
-EOF
-
-mkdir -p ${STACK_ROOT}/download-cache
-
-rm -rf All
-mkdir All
-cd All
-
-cat << EOF > stack.yaml
-resolver:            ${RESOLVER}
-EOF
-
-cat << EOF > All.cabal
-name:                All
-version:             0.1
-synopsis:            All
-description:         Please see README.md
-homepage:            All
-license:             PublicDomain
-author:              Automatic
-maintainer:          x@x.com
-copyright:           2016 Automatic
-category:            Console
-build-type:          Basic
-cabal-version:       >= 1.10
-
-executable all
-  hs-source-dirs:      .
-  main-is:             Main.hs
-  default-language:    Haskell2010
-  build-depends:       base
-EOF
-
-cat << EOF > Main.hs
-main = print "Hello"
-EOF
-
-cat << EOF > Setup.hs
-import           Distribution.Simple
-
-main :: IO ()
-main = defaultMain
-EOF
-
-stack setup
-stack list-build-plan | awk -F" " '{print "             , " $1}' >> All.cabal || exit -1
-stack build --dry-run --prefetch
-cd -
+stack --no-system-ghc setup --resolver=${resolver}
+stack unpack `stack --resolver=${resolver} list-build-plan` --fetch-only
 
 make_srpm() {
     cleanups () {
@@ -136,22 +94,21 @@ make_srpm() {
     RPM_TARGET_DIR=`mktemp --tmpdir -d XXXXXXrpm-packaging`
     mkdir -p ${RPM_TARGET_DIR}/{SPECS,SOURCES}
 
-    PKG_NAME=stackage-dist-${RESOLVER}
+    PKG_NAME=stackage-dist
     SPEC_FILE=${RPM_TARGET_DIR}/SPECS/${PKG_NAME}.spec
     PKG_VERSION=1
     PKG_RELEASE=1
 
     cat `which stack` | gzip -c > ${RPM_TARGET_DIR}/SOURCES/stack-bin.gz
-    cd ${STACK_ROOT}
-    tar -czf ${RPM_TARGET_DIR}/SOURCES/stack-root-download-cache.tar.gz download-cache
-    tar --exclude indices/Hackage/packages \
-	--exclude indices/Hackage/git-update \
-	-czf ${RPM_TARGET_DIR}/SOURCES/stack-root-indices.tar.gz indices
-    cp ${t}/helpers.sh ${RPM_TARGET_DIR}/SOURCES/helpers.sh
+    tar --exclude stack-root/programs/\*/\*.tar.xz \
+        -czf ${RPM_TARGET_DIR}/SOURCES/stack-root.tar.gz stack-root
+
+    echo ${STACK_ROOT} >> ${RPM_TARGET_DIR}/SOURCES/stack-root-path.txt
 
     cat ${t}/stackage-dist.rpm.spec \
 	| sed s/@@PKG_VERSION@@/${PKG_VERSION}/g \
 	| sed s/@@PKG_NAME@@/${PKG_NAME}/g \
+	| sed s/@@RESOLVER@@/${resolver}/g \
 	| sed s/@@PKG_RELEASE@@/${PKG_RELEASE}/g > ${SPEC_FILE}
 
     (rpmbuild -bs \
@@ -172,7 +129,7 @@ make_sdeb() {
 	rm -rf ${tempdir}
     }
 
-    PKG_NAME=stackage-dist-${RESOLVER}
+    PKG_NAME=stackage-dist-${resolver}
     PKG_VERSION=1.0
     PKG_RELEASE=1
     PKG_SITE=https://github.com/kernelim/stackage-dist
